refactor(search, recipe, template): centralize template search logic, update recipe deletion, and improve modal integration
- Centralized template search logic in application layer for better maintainability.
- Updated recipe deletion to return void and improved error handling.
- Refactored modal components to use new search and mutation patterns.
- Adjusted types and props for consistency across components.
- Updated tests and type checks to ensure all changes pass.
