+++ src/sections/search/components/TemplateSearchModal.tsx
@@ -1,7 +1,6 @@
 import {
   type Accessor,
   createEffect,
-  createResource,
   createSignal,
   type Setter,
   Show,
@@ -14,21 +13,9 @@ import {
   currentDayDiet,
   targetDay,
 } from '~/modules/diet/day-diet/application/dayDiet'
-import {
-  fetchFoodById,
-  fetchFoods,
-  fetchFoodsByName,
-} from '~/modules/diet/food/application/food'
-import { type Food } from '~/modules/diet/food/domain/food'
 import { type ItemGroup } from '~/modules/diet/item-group/domain/itemGroup'
 import { type MacroNutrients } from '~/modules/diet/macro-nutrients/domain/macroNutrients'
 import { getMacroTargetForDay } from '~/modules/diet/macro-target/application/macroTarget'
-import {
-  fetchRecipeById,
-  fetchUserRecipeByName,
-  fetchUserRecipes,
-} from '~/modules/diet/recipe/application/recipe'
-import { type Recipe } from '~/modules/diet/recipe/domain/recipe'
 import { type Template } from '~/modules/diet/template/domain/template'
 import { type TemplateItem } from '~/modules/diet/template-item/domain/templateItem'
 import {
@@ -37,19 +24,21 @@ import {
 } from '~/modules/diet/template-item/domain/templateItem'
 import {
   fetchRecentFoodByUserTypeAndReferenceId,
-  fetchUserRecentFoods,
   insertRecentFood,
   updateRecentFood,
 } from '~/modules/recent-food/application/recentFood'
 import { createNewRecentFood } from '~/modules/recent-food/domain/recentFood'
 import {
+  refetchTemplates,
+  setDebouncedSearch,
   setTemplateSearchTab,
+  templates,
   templateSearch,
   templateSearchTab,
 } from '~/modules/search/application/search'
 import { showSuccess } from '~/modules/toast/application/toastManager'
 import { showError } from '~/modules/toast/application/toastManager'
-import { currentUser, currentUserId } from '~/modules/user/application/user'
+import { currentUserId } from '~/modules/user/application/user'
 import { BarCodeButton } from '~/sections/common/components/BarCodeButton'
 import { Modal } from '~/sections/common/components/Modal'
 import { PageLoading } from '~/sections/common/components/PageLoading'
@@ -275,130 +264,6 @@ export function TemplateSearchModal(props: TemplateSearchModalProps) {
   )
 }
 
-const fetchRecentsForModal = async (): Promise<readonly Template[]> => {
-  const recentItems = await fetchUserRecentFoods(currentUserId())
-  const foodIds = recentItems
-    .filter((r) => r.type === 'food')
-    .map((r) => r.reference_id)
-  const recipeIds = recentItems
-    .filter((r) => r.type === 'recipe')
-    .map((r) => r.reference_id)
-
-  // Fetch foods and recipes in parallel
-  const foods =
-    foodIds.length > 0
-      ? await Promise.all(
-          foodIds.map((id) => fetchFoodById(id).catch(() => null)),
-        )
-      : []
-  const recipes =
-    recipeIds.length > 0
-      ? await Promise.all(
-          recipeIds.map((id) => fetchRecipeById(id).catch(() => null)),
-        )
-      : []
-
-  // Filter out nulls (not found)
-  const validFoods = (foods as (Food | null)[]).filter(
-    (f): f is Food => f !== null,
-  )
-  const validRecipes = (recipes as (Recipe | null)[]).filter(
-    (r): r is Recipe => r !== null,
-  )
-
-  // Sort by recency (last_used)
-  const templates: Template[] = []
-  for (const recent of recentItems) {
-    if (recent.type === 'food') {
-      const food = validFoods.find((f) => f.id === recent.reference_id)
-      if (food) templates.push(food)
-    } else {
-      const recipe = validRecipes.find((r) => r.id === recent.reference_id)
-      if (recipe) templates.push(recipe)
-    }
-  }
-  return templates
-}
-
-const fetchFoodsForModal = async (): Promise<readonly Food[]> => {
-  const getAllowedFoods = async () => {
-    switch (templateSearchTab()) {
-      case 'favorites':
-        return currentUser()?.favorite_foods ?? []
-      case 'recent':
-        return (await fetchUserRecentFoods(currentUserId())).map(
-          (food) => food.reference_id,
-        )
-      default:
-        return undefined // Allow any food
-    }
-  }
-
-  const limit =
-    templateSearchTab() === 'favorites'
-      ? undefined // Show all favorites
-      : 50 // Show 50 results
-
-  const allowedFoods = await getAllowedFoods()
-  console.debug('[TemplateSearchModal] fetchFunc', {
-    tab: templateSearchTab(),
-    search: templateSearch(),
-    limit,
-    allowedFoods,
-  })
-
-  let foods: readonly Food[]
-  if (templateSearch() === '') {
-    foods = await fetchFoods({ limit, allowedFoods })
-  } else {
-    foods = await fetchFoodsByName(templateSearch(), { limit, allowedFoods })
-  }
-
-  if (templateSearchTab() === 'recent') {
-    foods = (
-      await Promise.all(
-        allowedFoods?.map(async (foodId) => {
-          let food: Food | null = null
-          const alreadyFechedFood = foods.find((food) => food.id === foodId)
-          if (alreadyFechedFood === undefined) {
-            console.debug(
-              `[TemplateSearchModal] Food is not already fetched: ${foodId}`,
-            )
-            food = await fetchFoodById(foodId)
-          } else {
-            console.debug(
-              `[TemplateSearchModal] Food is already fetched: ${foodId}`,
-            )
-            food = alreadyFechedFood
-          }
-          return food
-        }) ?? [],
-      )
-    ).filter((food): food is Food => food !== null)
-  }
-  return foods
-}
-
-const fetchRecipes = async (): Promise<readonly Recipe[]> => {
-  if (templateSearch() === '') {
-    return await fetchUserRecipes(currentUserId())
-  } else {
-    return await fetchUserRecipeByName(currentUserId(), templateSearch())
-  }
-}
-
-const fetchFunc = async () => {
-  const tab_ = templateSearchTab()
-  switch (tab_) {
-    case 'recent':
-      return await fetchRecentsForModal()
-    case 'recipes':
-      return await fetchRecipes()
-    default:
-      return await fetchFoodsForModal()
-  }
-}
-
 export function TemplateSearch(props: {
   modalVisible: Accessor<boolean>
   barCodeModalVisible: Accessor<boolean>
@@ -412,25 +277,14 @@ export function TemplateSearch(props: {
   // TODO:   Determine if user is on desktop or mobile to set autofocus
   const isDesktop = false
 
-  const [debouncedSearch, setDebouncedSearch] = createSignal(templateSearch())
-
   const { typing, onTyped } = useTyping({
     delay: TYPING_TIMEOUT_MS,
     onTypingEnd: () => {
       setDebouncedSearch(templateSearch())
       console.debug(`[TemplateSearchModal] onTyped called`)
-      void refetch()
+      void refetchTemplates()
     },
   })
-
-  const [templates, { refetch }] = createResource(
-    () => ({
-      search: debouncedSearch(),
-      tab: templateSearchTab(),
-      userId: currentUserId(),
-    }),
-    fetchFunc,
-  )
 
   createEffect(() => {
     templateSearch()
@@ -478,10 +332,7 @@ export function TemplateSearch(props: {
           setItemEditModalVisible={props.setItemEditModalVisible}
           setSelectedTemplate={props.setSelectedTemplate}
           typing={typing}
-          // eslint-disable-next-line solid/reactivity
-          refetch={async () => {
-            await refetch()
-          }}
+          refetch={refetchTemplates}
         />
       </Suspense>
     </>
